[NFC] Refactor ValidatingSolver out of Solver.cpp into its own file ``ValidatingSolver.cpp``. Whilst I'm here also clang-format the modified code.
